-Changed findValidMoves arraylist type to short from integer(idk if its meaningful but I did). -StartGame method was only accepting player1 inputs, added CurrentPlayer instead. -For computer to make a move it needs to know validMoves so added validMovesList arg to getInput method of player class. -Small UI changes.

+++ Reversi/ReversiGUI.java
@@ -231,13 +231,13 @@ public class ReversiGUI {
         frame.revalidate();
     }
 
-    public void addValidMoves(ArrayList<Integer> validMovesList) {
+    public void addValidMoves(ArrayList<Short> validMovesList) {
         for (int index : validMovesList) {
             squares[index].setState(MySquare.GUESS);
         }
     }
 
-    public void removeValidMoves(ArrayList<Integer> validMovesList) {
+    public void removeValidMoves(ArrayList<Short> validMovesList) {
         for (int index : validMovesList) {
             squares[index].setState(Board.EMPTY);
         }
